fix(docker-driver): Multiplex the client.batchSend
promtail `client` stop is not propaged properly to the its `batchSend` method
in the `run` method.

though `run` method multiplex on `c.quit` channel, the other branch can block
in `batchSend` without even listening on the `c.quit` channel on the `run` method.
This fixes the it by multiplex on `c.quit` inside the batchSend as well.

Other suggestion: expose `client.run` to promtail client and they control it via
context. But that requires API changes like accepting `ctx` in `client.Run` method.